+++ src/features/auth/ui/LoginForm/LoginForm.tsx
@@ -1,23 +1,19 @@
 'use client';
 
-import { useRouter } from 'next/navigation';
 import { useState } from 'react';
 
-import { LoginFormValues } from '@/shared/schema/auth';
-import { useModalStore } from '@/shared/stores/useModalStore';
 import Button from '@/shared/ui/Button/Button';
 import Input from '@/shared/ui/Input/Input';
 import Label from '@/shared/ui/Label';
 
 import { useLoginForm } from '../../model/useLoginForm';
+import { useLoginSubmit } from '../../model/useLoginSubmit';
 
 import VisibleButton from './VisibleButton';
 
 const LoginForm = () => {
-  const router = useRouter();
   const [isVisible, setIsVisible] = useState(false);
   const [isCapsLockOn, setIsCapsLockOn] = useState(false);
-  const { openAlert } = useModalStore();
 
   const {
     register,
@@ -25,27 +21,7 @@ const LoginForm = () => {
     formState: { errors, isValid, isSubmitting },
   } = useLoginForm();
 
-  // **추후 수정** 로그인 api 연결
-  const handleLogin = async (data: LoginFormValues) => {
-    try {
-      const res = await fetch('/api/login', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(data),
-      });
-
-      if (!res.ok) {
-        // 로그인 실패 모달 띄움 에러 메세지랑
-        openAlert('로그인에 실패했습니다.');
-        return;
-      }
-
-      router.push('/');
-    } catch (e) {
-      console.log('네트워크 에러', e);
-      openAlert('네트워크 연결 상태를 확인해주세요.');
-    }
-  };
+  const handleLogin = useLoginSubmit();
 
   const handleVisiblity = () => {
     setIsVisible(!isVisible);
